+++ python/tests/conftest.py
@@ -106,3 +106,13 @@ def topphatt_bit_sliced_weight_snapshot() -> dict:
         data = json.load(file)
 
     return data
+
+@fixture(scope="module")
+def hubbard_weight_snapshot() -> dict:
+    folder = Path(__file__).parent
+    filename = "./data/hubbard_weight_snapshot.json"
+
+    with open(folder.joinpath(filename), "rb") as file:
+        data = json.load(file)
+
+    return data